move components to components folder and update dependencies
