+++ src/components/Login/Login.jsx
@@ -3,7 +3,7 @@ import './Login.css';
 import {useEffect, useState} from "react";
 import {useNavigate} from "react-router-dom";
 import axios from "axios";
-import {setToken} from "../helpers/storage";
+import {setToken} from "../../helpers/storage";
 
 export function Login({isLoggedIn, setIsLoggedIn}) {
     const [email, setEmail] = useState("");
